requirements.txt
azure-functions
azure-storage-blob
opencv-python-headless
requests
python-dotenv
numpy
